Move isSuperUser function to auth lib
