+++ lib/auth.js
@@ -14,14 +14,14 @@ export function removeAuthCookie() {
   cookie.remove(accessTokenId)
 }
 
-export function auth(ctx = {}) {
-  let cookie
-  if (ctx.req) {
-    const cookieHeader = ctx.req.headers && ctx.req.headers.cookie
-    cookie = new Cookies(cookieHeader)
-  } else {
-    cookie = new Cookies()
-  }
+export function isSuperUser(user) {
+  return user?.id === process.env.SUPER_USER
+}
+
+export function auth(ctx) {
+  // Get cookies from request headers if executing on server side, or document.cookies
+  // if running on the client
+  const cookie = new Cookies(ctx?.req?.headers?.cookie)
 
   const token = cookie.get(accessTokenId)
   if (!token) {
