+++ lib/utils.js
@@ -9,15 +9,6 @@ import isObject from 'lodash/isObject'
 /**
  * Get the human readable version of a group or profile id
  *
- * @param {object} user - user to check
- */
-export function isSuperUser(user) {
-  return user?.id === process.env.SUPER_USER
-}
-
-/**
- * Get the human readable version of a group or profile id
- *
  * @param {string} id - group or profile id to convert
  * @param {boolean} [onlyLast] - don't include Paper# token in output
  */
